// front-end/e2e/support/_shared/interfaces/sidebar.interface.ts
export interface Sidebar {
  logo: string;
  access: {
    btnEntrar: string;
    btnCriarConta: string;
    btnSair: string;
    modalConfirmLogout: string;
    avatar: string;
  };
  menu: {
    search: string;
    nav: {
      home: string;
      fornecedores: string;
      produtos: string;
    };
  };
}